pyproject.toml
[project]
name = "hydra-htcondor-launcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "htcondor>=25.1.0",
    "hydra-core>=1.3.2",
]
